Update and rename ModalHeaderSpec.js to ModalHeaderSpec.tsx
